backend/requirements.txt
fastapi
uvicorn
numpy
tensorflow
scikit-learn
pillow 
python-multipart